pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relational-knowledge-graph"
version = "0.1.0"
description = "Database schema knowledge graph for LLM-assisted SQL generation"
authors = [
    {name = "Adrian", email = ""}
]
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.1.0",
    "networkx>=3.2",
    "plotly>=5.17.0",
    "matplotlib>=3.8.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "pandas>=2.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
rkg = "relational_kg.cli:main"